modified core structure
